refactor(cli): decompose tui/state.rs into 5 focused module files
Earlier 'tui.rs split' (PR #66) was mostly a rename plus extracting
run_tui/handle_tui_command from main.rs. The actual god file
(state.rs at 2226 lines) was untouched. This PR decomposes its
internals.

state.rs goes from 2226 -> 1349 lines (-877, -39%).

Five new files in apps/netscli-cli/src/tui/, each with a single
responsibility:

  palette.rs           70 lines  Palette struct + palette() lookup +
                                 is_light_terminal() detection via
                                 $COLORFGBG.

  command_catalog.rs  200 lines  CommandDef descriptor + the
                                 platform-conditional COMMAND_DEFS
                                 arrays (with/without /pcap entry
                                 depending on the pcap feature) +
                                 help_lines() renderer.

  history.rs           21 lines  Public EntryState + HistoryEntry
                                 types used by the TuiApp scrollback
                                 and tui_export.rs.

  config.rs            89 lines  ConfigItemKind + ConfigUiState +
                                 its impl (new/item/len/
                                 clamp_selected/ensure_visible) +
                                 the top-level UiMode enum.

  widgets.rs          582 lines  Generic rendering helpers
                                 (configure_input, label_style,
                                 line_with_drawn_cursor,
                                 build_suggestion_line,
                                 truncate_styled_spans, boxed_lines,
                                 gradient_text_line_with_left_pad,
                                 scrollbar_thumb, cycle_option,
                                 cycle_unit, config_line, etc.)
                                 plus the INPUT_PLACEHOLDER constant.

state.rs now contains only the TuiApp struct and its impl block
(plus the integration tests that drive TuiApp at the rendering
level — those couldn't move to widgets.rs because they reference
TuiApp's private state).

mod.rs's public re-exports are reorganized:
  pub use command_catalog::help_lines;
  pub use history::{EntryState, HistoryEntry};
  pub use state::TuiApp;

External call sites (main.rs, tui_export.rs) keep their
'use crate::tui::Foo' imports working unchanged.

Verified locally:
- cargo check -p netscli clean
- cargo clippy -p netscli --all-targets -- -D warnings clean
- cargo fmt --all -- --check clean
- cargo test -p netscli — all 25 tests pass
